Combined gym envs into one. Minor code cleanup/refactoring. Currently fails gym tests!
