+++ examples/ppo_hover_train.py
@@ -1,28 +1,15 @@
-import gymnasium as gym
-import numpy as np
-import matplotlib.pyplot as plt
 import os
 from datetime import datetime
+import torch
 
 from rotorpy.vehicles.crazyflie_params import quad_params  # Import quad params for the quadrotor environment.
 
 # Import the QuadrotorEnv gymnasium environment using the following command.
-from rotorpy.learning.quadrotor_environments import QuadrotorEnv
+from rotorpy.learning.quadrotor_environments import QuadrotorEnv, make_default_vec_env
+from rotorpy.learning.quadrotor_reward_functions import vec_hover_reward, vec_hover_reward_positive
+from rotorpy.learning.learning_utils import *
 
-# Reward functions can be specified by the user, or we can import from existing reward functions.
-from rotorpy.learning.quadrotor_reward_functions import hover_reward, hover_reward_positive
 
-"""
-In this script, we demonstrate how to train a hovering control policy in RotorPy using Proximal Policy Optimization. 
-We use our custom quadrotor environment for Gymnasium along with stable baselines for the PPO implementation. 
-
-The task is for the quadrotor to stabilize to hover at the origin when starting at a random position nearby. 
-
-Training can be tracked using tensorboard, e.g. tensorboard --logdir=<log_dir>
-
-"""
-
-# First we'll set up some directories for saving the policy and logs.
 models_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "rotorpy", "learning", "policies")
 log_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "rotorpy", "learning", "logs")
 if not os.path.exists(models_dir):
@@ -34,50 +21,42 @@ if not os.path.exists(log_dir):
 try:
     import stable_baselines3
 except:
-    raise ImportError('To run this example you must have Stable Baselines installed via pip install stable_baselines3')
+    raise ImportError('To run this example you must have Stable Baselines installed. Try installing rotorpy with pip install rotorpy.[learning]')
 
 from stable_baselines3 import PPO                                   # We'll use PPO for training.
 from stable_baselines3.ppo.policies import MlpPolicy                # The policy will be represented by an MLP
+from stable_baselines3.common.vec_env import VecMonitor
+from stable_baselines3.common.callbacks import EvalCallback, CheckpointCallback, CallbackList
 
-num_cpu = 4   # for parallelization
-
-# Choose the weights for our reward function. Here we are creating a lambda function over hover_reward.
-# reward_function = lambda obs, act: hover_reward(obs, act, weights={'x': 1, 'v': 0.1, 'w': 0, 'u': 1e-5})
-reward_function = hover_reward_positive
-
-# Make the environment. For this demo we'll train a policy to command collective thrust and body rates.
-# Turning render_mode="None" will make the training run much faster, as visualization is a current bottleneck. 
-env = gym.make("Quadrotor-v0", 
-                control_mode ='cmd_motor_speeds', 
-                reward_fn = reward_function,
-                quad_params = quad_params,
-                max_time = 5,
-                world = None,
-                sim_rate = 100,
-                render_mode='None')
-
-# from stable_baselines3.common.env_checker import check_env
-# check_env(env, warn=True)  # you can check the environment using built-in tools
-
-# Reset the environment
-observation, info = env.reset(options={'initial_state': 'random', 'pos_bound': 2, 'vel_bound': 0})
-
-# Create a new model
-model = PPO(MlpPolicy, env, n_steps=128, verbose=1, ent_coef=0.01, tensorboard_log=log_dir)
-
-# Training... 
-num_timesteps = 20_000
-num_epochs = 10
+control_mode = "cmd_ctatt"
+num_envs = 512
+device = torch.device("cpu")
+random_quad_params = generate_random_vehicle_params(num_envs, device, quad_params, crazyflie_randomizations)
+env = make_default_vec_env(num_envs,
+                           random_quad_params,
+                           control_mode,
+                           device,
+                           render_mode="None",
+                           reward_fn=vec_hover_reward_positive)
+env.reset_options["params"] = "random"
+env.reset_options["vel_bound"] = 0.2
+wrapped_env = VecMonitor(env)
 
 start_time = datetime.now()
 
-epoch_count = 0
-while True:  # Run indefinitely..
+eval_env = VecMonitor(make_default_vec_env(10, quad_params, control_mode, device, render_mode="None", reward_fn=vec_hover_reward_positive))
+checkpoint_callback = CheckpointCallback(save_freq=max(50000//num_envs, 1), save_path=f"{models_dir}/PPO/hover_{control_mode}_{start_time.strftime('%b-%d_%H-%M')}/",
+                                         name_prefix='hover')
+eval_callback = EvalCallback(eval_env, eval_freq=1e6//num_envs, deterministic=True, render=False)
+model = PPO(MlpPolicy,
+            wrapped_env,
+            n_steps=32,
+            batch_size=1024,
+            verbose=1,
+            device=device,
+            tensorboard_log=log_dir)
 
-    # This line will run num_timesteps for training and log the results every so often.
-    model.learn(total_timesteps=num_timesteps, reset_num_timesteps=False, tb_log_name="PPO-Quad_cmd-motor_"+start_time.strftime('%H-%M-%S'))
-
-    # Save the model
-    model.save(f"{models_dir}/PPO/{start_time.strftime('%H-%M-%S')}/hover_{num_timesteps*(epoch_count+1)}")
-
-    epoch_count += 1
+num_timesteps = 10e6
+model.learn(total_timesteps=num_timesteps, reset_num_timesteps=False,
+            tb_log_name="PPO-QuadVec_"+control_mode+"_"+ start_time.strftime('%b-%d-%H-%M'),
+            callback=CallbackList([checkpoint_callback, eval_callback]))
